added condition to check if guess was already submitted
