+++ TextTwist/WordData.swift
@@ -1,0 +1,30 @@
+//
+//  WordData.swift
+//  TextTwist
+//
+//  Created by Alexander George Legaspi on 8/6/19.
+//  Copyright © 2019 C4Q . All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+class WordData {
+    static let allInfo: [TestTwistInfo] =
+        
+        [
+            
+            TestTwistInfo(wordCount: 27,
+                          letters: "cdenno",
+                          words: ["conned", "coden","coned", "donne", "nonce","code","coed","cone","conn","deco","done","neon","node","none","once", "cod","con","den","doc","doe","don","end","eon","nod","ode","one"]),
+            
+            TestTwistInfo(wordCount: 18,
+                          letters: "ehhssu",
+                          words: ["hushes", "shush", "hehs","hues","hush","shes","sues","uses", "ess","heh","hes","hue","huh","she","shh","sue","use"]),
+            
+            TestTwistInfo(wordCount: 14,
+                          letters: "hisswy",
+                          words: ["swishy", "hissy","swish", "hiss","whys","wish","wiss","ywis", "his","shy","sis","why","wis"])
+    ]
+    
+}
